refactor: 클럽 아이템 ui model 재활용 가능하도록 수정
