+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/mapper/ClubMapper.kt
@@ -1,11 +1,11 @@
 package com.happy.friendogly.presentation.ui.club.mapper
 
 import com.happy.friendogly.domain.model.Club
-import com.happy.friendogly.presentation.ui.club.list.ClubListUiModel
+import com.happy.friendogly.presentation.ui.club.common.model.ClubItemUiModel
 import com.happy.friendogly.presentation.ui.club.common.model.ClubPet
 
-fun Club.toPresentation(): ClubListUiModel {
-    return ClubListUiModel(
+fun Club.toPresentation(): ClubItemUiModel {
+    return ClubItemUiModel(
         clubId = id,
         clubPoster = imageUrl?.ifEmpty { null },
         maximumNumberOfPeople = memberCapacity,
